The problem of distinguishing prime numbers from composite numbers and of resolving the latter into their prime factors is known to be one of the most important and useful in arithmetic

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -8,11 +8,11 @@
 
 int main(void)
 {
-	int factor;
-	int largest;
-	int prime;
-	int count;
-	int n;
+	long factor;
+	long largest;
+	long prime;
+	long count;
+	long n;
 
 	long number = 612852475143;
 
@@ -36,7 +36,7 @@ int main(void)
 		}
 	}
 
-	printf("%d\n", largest);
+	printf("%ld\n", largest);
 
 	return (0);
 }
